Разделите код, уже написанный в main.js, на отдельные ES2015-модули
